refactor directory & import statements
