+++ src/app/layout.tsx
@@ -1,15 +1,12 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import {
-  ClerkProvider,
-  SignedIn,
-  SignedOut,
-  SignInButton,
-  UserButton
+  ClerkProvider
 } from "@clerk/nextjs";
 
-import "./globals.css";
 import NavBar from "@/components/main/NavBar";
+
+import "./globals.css";
 
 const inter = Inter({ subsets: ["latin"] });
